This patch is all about supporting IPv6 extension headers better.
Before this patch, all extension headers in the system were left
somewhat compressed-- 6lowpan specifies that their length is stored in
units of single octets instead of 8-octet units like they are are in
full IPv6 networks.

In this patch, I change it so the represention of IPv6 packets once
unpacked is identical to that of IPv6 packets anywhere else -- this
means that extension header lengths are stored as 8-octet units.  The
6lowpan compressor will now check for and remove any trailing Pad1 or
PadN TLV options when compressing HBH or Destination option extension
headers.

This means that the interface to extension headers has changed, and
will require modification of any applications who were doing this.
This change was necessary because when packets are routed between an
external network and a RPL domain, the headers would have needed to
have had to be modified or removed, and this seemed like the most
principled way of doing this.

Finally, draft-hui-6man-rpl-headers-00 suggests that it may be
desirable to insert and remove headers instead of re-encapsulating
packets in IP-in-IP encapsulation when additional headers are
necessary.  For us, this is necessary when packets containing a RPL
HBH TLV option leave the network.  The proposed TLV option type is
0x6b, which indicates that the packet should be dropped if the option
is not recognized and we are required to remove this option or
reencapsulate the packet when the packet leave the RPL domain.
Therefore, PppRouter has been modified convert any RPL TLV options
inside of HBH headers to PadN options when the packet is being sent
over the PPP interface.



git-svn-id: https://tinyos-main.googlecode.com/svn/trunk@5631 285f258d-8064-b571-fbc6-9656dc998e90